extract component lessonEditor
